requirements.txt
numpy==2.1.0
pandas==2.2.2
PyYAML==6.0.2
scikit_learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
skimage==0.0
tensorboardX==2.6.2.2
torch==2.3.1
torchnet==0.0.4
torchvision==0.18.1
